CBL-3349 : getPendingDocIDs and isDocumentPending to have collection as argument (#1505)
CBL-3087: Correct bug in array_count that counts null values (#1507)
CBL-3333: Use size_t for types representing array size (#1506)
CBL-3306: Fix bug where expiration only works on default collection (#1503)
CBL-3298: Make c4db_hasScope multi handle aware (#1502)
CBL-3057: Set error information to zero on some c4repl methods (#1501)
Set the create flag when rekeying (#1489)
CBL-2874: Correct an outdated documentation comment (#1500)

:warning: API change to c4repl_getPendingDocIDs: added CollectionSpec to parameters
:warning: API change to c4repl_isDocumentPending: added CollectionSpec to paramters

+++ C/include/c4ReplicatorTypes.h
@@ -226,7 +226,7 @@ C4API_BEGIN_DECLS
         // If collections == nullptr, we will use the deprecated fields to build
         // the internal config for one collection being the default collection.
         C4ReplicationCollection             *collections;
-        unsigned                            collectionCount;
+        size_t                               collectionCount;
     } C4ReplicatorParameters;
